s390/smp: fix memblock_phys_free() vs memblock_free() confusion
Conflicts:
	arch/s390/kernel/smp.c - The backport of
		3ecc68349bba ("memblock: rename memblock_free to memblock_phys_free")
		dropped the change to this file because it didn't auto-merge,
		and causes another merge conflict with this patch. Just make
		the change so the end result is the same.
	arch/s390/kernel/setup.c - The backport of 3ecc68349bba was also
		supposed to change this file but didn't. It was fixed up in
		merge commit
		0b707e572a19 ("Merge tag 's390-5.16-1' of git://git.kernel.org/pub/scm/linux/kernel/git/s390/linux")
		right that wrong here as well.

Bugzilla: https://bugzilla.redhat.com/2120352

commit fcfcba6dfc9a57da9c816667c20614ddfd4def10
Author: Heiko Carstens <hca@linux.ibm.com>
Date:   Mon Dec 13 19:00:57 2021 +0100

    s390/smp: fix memblock_phys_free() vs memblock_free() confusion

    memblock_phys_free() is used on a virtual address. Fix this by using
    memblock_free().

    Note: this doesn't fix a bug currently, since virtual and physical
    addresses are identical.

+++ arch/s390/kernel/smp.c
@@ -861,7 +861,7 @@ void __init smp_detect_cpus(void)
 
 	/* Add CPUs present at boot */
 	__smp_rescan_cpus(info, true);
-	memblock_free((unsigned long)info, sizeof(*info));
+	memblock_free(info, sizeof(*info));
 }
 
 static void smp_init_secondary(void)
